3.1.12 Klasa Node nie musi byc publiczna, klasy factory musza byc publiczne aby mozna bylo z nich korzystac, z klas FIFORyhmer, HanoiRyhmer, DefaultCountingOutRyhmer korzystaja inne klasy.
