+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -55,3 +55,30 @@ public class IntLinkedList {
     }
 
 }
+
+class Node {
+
+    int value; // dostep do zmiennej odbywa sie poprzez funkcje get i set
+    Node prev; // dostep do zmiennej odbywa sie poprzez funkcje get i set
+    Node next; // dostep do zmiennej odbywa sie poprzez funkcje get i set
+    //
+    Node(int i) {
+        setValue(i);
+    }
+
+    public int getValue() {
+        return value;
+    }
+
+    public void setValue(final int value) {
+        this.value = value;
+    }
+
+    public Node getPrev() {
+        return prev;
+    }
+
+    public Node getNext() {
+        return next;
+    }
+}
